@@ -18,22 +18,28 @@
     }
   ],
   "scripts": {
-    "start": "node app.js",
-    "postinstall": "./postinstall.sh"
+    "start": "node index.js",
+    "postinstall": "./postinstall.sh",
+    "test": "NODE_ENV=test BLUEBIRD_DEBUG=1 jest"
   },
   "dependencies": {
     "async": "^2.0.0",
-    "babel-preset-react": "^6.11.1",
+    "body-parser": "^1.16.1",
     "bootstrap": "^3.3.6",
     "bower": "^1.7.9",
     "browserify": "^13.0.1",
-    "connect-redis": "~1.4.7",
+    "compression": "^1.6.2",
+    "connect-redis": "^3.2.0",
     "consolidate": "^0.14.1",
+    "cookie-parser": "^1.4.3",
+    "csurf": "^1.9.0",
+    "d3-queue": "^3.0.3",
     "elasticsearchclient": "~0.5.3",
     "envify": "~1.2.1",
     "es5-shim": "^4.5.9",
-    "express": "^3.21.2",
-    "express-expose": "0.3.3",
+    "express": "^4.14.1",
+    "express-expose": "^0.3.4",
+    "express-session": "^1.15.1",
     "geopoint": "^1.0.1",
     "griddle-react": "^0.6.1",
     "hamljs": "^0.6.2",
@@ -41,13 +47,10 @@
     "http-proxy": "~1.8.1",
     "jquery": "~2.1.4",
     "jquery-ui-browserify": "1.11.0-pre-seelio",
-    "leaflet": "0.7.3",
-    "leaflet-image": "^0.2.1",
-    "leaflet-sleep": "^0.5.0",
     "less": "^2.7.1",
     "lodash": "^4.13.1",
-    "node-markdown": "^0.1.1",
-    "node-uuid": "^1.4.7",
+    "method-override": "^2.3.7",
+    "morgan": "^1.8.1",
     "openid": "^2.0.1",
     "pg": "^4.5.6",
     "pouchdb": "^5.4.5",
@@ -58,19 +61,31 @@
     "reactify": "^1.1.1",
     "redis": "^2.6.2",
     "request": "^2.73.0",
+    "serve-favicon": "^2.4.0",
+    "serve-static": "^1.11.2",
     "swig": "^1.4.2",
     "uglify-js": "~2.4.24",
-    "underscore": "^1.8.3"
+    "underscore": "^1.8.3",
+    "winston": "^2.3.1"
   },
   "devDependencies": {
     "assert": "~1.3.0",
-    "babel": "^5.0",
-    "babel-preset-es2015": "^6.9.0",
+    "babel-cli": "^6.18.0",
+    "babel-core": "^6.23.1",
+    "babel-env": "^1.3.1",
+    "babel-jest": "^18.0.0",
+    "babel-plugin-module-resolver": "^2.5.0",
+    "babel-plugin-transform-promise-to-bluebird": "^1.1.1",
+    "babel-preset-es2015": "^6.22.0",
+    "babel-preset-node6": "^11.0.0",
     "babel-preset-react": "^6.11.1",
+    "babel-preset-stage-3": "^6.17.0",
+    "babel-register": "^6.18.0",
     "babelify": "^7.3.0",
-    "chai": "^3.5.0",
-    "docco": "~0.7.0",
     "es5-shim": "^4.5.9",
+    "eslint": "^3.16.0",
+    "eslint-import-resolver-babel-module": "^3.0.0",
+    "eslint-plugin-import": "^2.2.0",
     "gulp": "^3.9.0",
     "gulp-babel": "^6.1.2",
     "gulp-browserify": "^0.5.1",
@@ -81,20 +96,29 @@
     "gulp-rename": "^1.2.0",
     "gulp-uglify": "^1.4.2",
     "gulp-util": "^3.0.7",
-    "lorem": "~0.4.0",
-    "mocha": "^2.5.3",
+    "jest": "^18.0.0",
     "npm-shrinkwrap": "200.5.1",
+    "react-test-renderer": "^15.4.2",
     "reactify": "^1.1.1",
+    "supertest": "^3.0.0",
     "uglify-js": "~2.4.24",
     "vinyl-buffer": "^1.0.0",
     "vinyl-source-stream": "^1.1.0",
-    "watchify": "^3.7.0",
-    "zombie": "~2.5.1"
+    "watchify": "^3.7.0"
   },
   "keywords": [],
   "repository": "git://github.com/idigbio/idigbio-portal",
   "license": "LGPL-3.0",
   "engines": {
-    "node": "~0.10.46"
+    "node": ">=4"
+  },
+  "jest": {
+    "testPathIgnorePatterns": [
+      ".eslintrc.js",
+      "/node_modules/",
+      "init.js",
+      "/public/"
+    ],
+    "setupTestFrameworkScriptFile": "<rootDir>/__tests__/init"
   }
 }
