@@ -1,123 +1,130 @@
 var request = require('request');
 var async = require('async');
-var path = require('path'),
-    appDir = path.dirname(require.main.filename);
-var config = require(appDir+'/config/config');
-var RecordsetPage = require(appDir+'/public/client/js/react/build/recordset');
-var fields = require(appDir+'/public/client/js/lib/fields');
-//var RecordsetPage = require(appDir+'/public/react/build/recordset');
-module.exports = function(app, config) {
-    return {
-        collections: function(req,res){
+var RecordsetPage = require('public/client/js/react/build/recordset');
 
-            request.get({"url": 'http://idigbio.github.io/idb-us-collections/collections.json', "json": true}, function(err, resp, body){
-                res.render('collections', {
-                    activemenu: 'publishers',
-                    user: req.user,
-                    token: req.session._csrf,
-                    data: JSON.stringify(body)
-                });
-            }); 
-        },
-        collection: function(req,res){
-            request.get({"url": 'http://idigbio.github.io/idb-us-collections/collections/'+req.params.id, "json": true}, function(err, resp, body){
-                
-                res.render('collection', {
-                    activemenu: 'publishers',
-                    user: req.user,
-                    token: req.session._csrf,
-                    data: JSON.stringify(body)
-                });
+import config from 'config/config'; // eslint-disable-line no-unused-vars
+import logger from 'app/logging'; // eslint-disable-line no-unused-vars
+
+// var RecordsetPage = require(appDir+'/public/react/build/recordset');
+export default {
+  collections: function(req, res) {
+
+    request.get({"url": 'http://idigbio.github.io/idb-us-collections/collections.json', "json": true}, function(err, resp, body) {
+      if(err) {
+        logger.error(err);
+      }
+      res.render('collections', {
+        activemenu: 'publishers',
+        user: req.user,
+        token: req.session._csrf,
+        data: JSON.stringify(body)
+      });
+    });
+  },
+  collection: function(req, res) {
+    request.get({"url": 'http://idigbio.github.io/idb-us-collections/collections/' + req.params.id, "json": true}, function(err, resp, body) {
+      if(err) {
+        logger.error(err);
+      }
+      res.render('collection', {
+        activemenu: 'publishers',
+        user: req.user,
+        token: req.session._csrf,
+        data: JSON.stringify(body)
+      });
+    });
+  },
+  publishers: function(req, res) {
+    res.render('publishers', {
+      activemenu: 'publishers',
+      user: req.user,
+      token: req.session._csrf
+    });
+  },
+  recordset: function(req, res) {
+    var flags = {};
+    var stotal = 0, mtotal = 0;
+    var rbody = {}, use = {};
+    var lastRecord = '', lastMedia = '';
+    request.get({"url": config.api + 'view/recordsets/' + req.params.id, "json": true}, function(err, resp, body) {
+      if(err) {
+        logger.error(err);
+      }
+      if(body.found === false) {
+        res.status(404);
+        res.render('404', {
+          activemenu: 'publishers',
+          user: req.user,
+          token: req.session._csrf,
+          id: req.params.id
+        });
+      } else {
+        rbody = body;
+        async.parallel([
+          function(cback) {
+            var q = {top_fields: ["flags"], count: 100, rq: {"recordset": req.params.id}};
+            request.post({"url": config.api + 'summary/top/records/', "json": true, "body": q}, function(a_err, a_resp, a_body) {
+              flags = a_body.flags;
+              cback(a_err, 'one');
             });
-        },
-        publishers: function(req, res) {
-            res.render('publishers', {
-                activemenu: 'publishers',
-                user: req.user,
-                token: req.session._csrf
+          },
+          function(cback) {
+            request.post({"url": config.api + 'summary/count/records/', "json": true, "body": {rq: {recordset: req.params.id}}}, function(a_err, a_resp, a_body) {
+              stotal = a_body.itemCount;
+              cback(a_err, 'two');
             });
-        },
-        recordset: function(req,res){
-            var flags = {};
-            var stotal=0,mtotal=0;
-            var rset={},rbody={},use;
-            var lastRecord='',lastMedia='';
-            var keys = Object.keys(fields.byDataTerm);
-            request.get({"url": config.api+'view/recordsets/'+req.params.id, "json": true}, function(err, resp, body){
-                if(body.found===false){
-                    res.status(404);
-                    res.render('404',{
-                        activemenu: 'publishers',
-                        user: req.user,
-                        token: req.session._csrf,
-                        id: req.params.id
-                    });                    
-                }else{
-                    rbody = body;
-                    rset = body.data;
-                    async.parallel([
-                        function(cback){
-                            var q ={top_fields:["flags"],count: 100, rq: {"recordset": req.params.id}};
-                            request.post({"url": config.api+'summary/top/records/', "json": true, "body": q}, function(err, resp, body){
-                                flags = body.flags;
-                                cback(null,'one');            
-                            });                
-                        },
-                        function(cback){
-                            request.post({"url": config.api+'summary/count/records/', "json": true, "body": {rq: {recordset: req.params.id}}}, function(err, resp, body){
-                                stotal = body.itemCount;
-                                cback(null,'two')                            
-                            });
-                        },
-                        function(cback){
-                            request.post({"url": config.api+'summary/count/media/', "json": true, "body": {mq: {recordset: req.params.id}}}, function(err, resp, body){
-                                mtotal = body.itemCount;
-                                cback(null,'three')                            
-                            });
-                        },
-                        function(cback){
-                            var params={"dateInterval":"month","recordset": req.params.id,"minDate":"2015-01-15"};
-                            request.post({"url": config.api+'summary/stats/search', "json": true, "body": params},function(err,resp,body){
-                                use=body;
-                                cback(null,'four');
-                            })
-                        },
-                        function(cback){
-                            request.post({"url": config.api+'summary/modified/records/', "json": true, "body": {rq: {recordset: req.params.id}}}, function(err, resp, body){
-                                lastRecord=body.lastModified.substring(0,10);
-                                cback(null,'five');
-                            })
-                        },
-                        function(cback){
-                            request.post({"url": config.api+'summary/modified/media/', "json": true, "body": {mq: {recordset: req.params.id}}}, function(err, resp, body){
-                                lastMedia=body.lastModified.substring(0,10);
-                                cback(null,'six');
-                            })
-                        }
-                    ],function(err,results){
-                        var React = require('react');
-                        var ReactDOMServer = require('react-dom/server');
-                        var Rp = React.createFactory(RecordsetPage);
-                        var lastmodified=lastRecord>=lastMedia?lastRecord:lastMedia;
-                        res.render('recordset',{
-                            activemenu: 'publishers',
-                            user: req.user,
-                            token: req.session._csrf,
-                            uuid: "'"+req.params.id+"'",
-                            mtotal: mtotal,
-                            stotal: stotal,
-                            lastmodified: lastmodified,
-                            flags: JSON.stringify(flags),
-                            recordset: JSON.stringify(rbody),
-                            use: JSON.stringify(use),
-                            content: ReactDOMServer.renderToString(Rp({mtotal: mtotal, stotal: stotal, flags: flags, recordset: rbody, use: use, lastmodified: lastmodified, uuid: req.params.id}))
-                        });
-                    })                           
-                }
+          },
+          function(cback) {
+            request.post({"url": config.api + 'summary/count/media/', "json": true, "body": {mq: {recordset: req.params.id}}}, function(a_err, a_resp, a_body) {
+              mtotal = a_body.itemCount;
+              cback(a_err, 'three');
             });
-        },
-        recordsetRedirect: function(req,res){
-            res.redirect('/portal/recordsets/'+req.params.id);
-        } 
-    }
-}
+          },
+          function(cback) {
+            var params = {"dateInterval": "month", "recordset": req.params.id, "minDate": "2015-01-15"};
+            request.post({"url": config.api + 'summary/stats/search', "json": true, "body": params}, function(a_err, a_resp, a_body) {
+              use = a_body;
+              cback(a_err, 'four');
+            });
+          },
+          function(cback) {
+            request.post({"url": config.api + 'summary/modified/records/', "json": true, "body": {rq: {recordset: req.params.id}}}, function(a_err, a_resp, a_body) {
+              lastRecord = a_body.lastModified.substring(0, 10);
+              cback(a_err, 'five');
+            });
+          },
+          function(cback) {
+            request.post({"url": config.api + 'summary/modified/media/', "json": true, "body": {mq: {recordset: req.params.id}}}, function(a_err, a_resp, a_body) {
+              lastMedia = a_body.lastModified.substring(0, 10);
+              cback(a_err, 'six');
+            });
+          }
+        ], function(a_err, results) {
+          if(a_err) {
+            logger.error(a_err); // This should probably actually be handled better.
+          }
+          var React = require('react');
+          var ReactDOMServer = require('react-dom/server');
+          var rp = React.createFactory(RecordsetPage);
+          var lastmodified = lastRecord >= lastMedia ? lastRecord : lastMedia;
+          res.render('recordset', {
+            activemenu: 'publishers',
+            user: req.user,
+            token: req.session._csrf,
+            uuid: "'" + req.params.id + "'",
+            mtotal: mtotal,
+            stotal: stotal,
+            lastmodified: lastmodified,
+            flags: JSON.stringify(flags),
+            recordset: JSON.stringify(rbody),
+            use: JSON.stringify(use),
+            content: ReactDOMServer.renderToString(rp({mtotal: mtotal, stotal: stotal, flags: flags, recordset: rbody, use: use, lastmodified: lastmodified, uuid: req.params.id}))
+          });
+        });
+      }
+    });
+  },
+  recordsetRedirect: function(req, res) {
+    res.redirect('/portal/recordsets/' + req.params.id);
+  }
+};
